+++ History.md
@@ -1,3 +1,8 @@
+unreleased
+==========
+
+  * Support array of fields to set
+
 0.0.0 / 2014-06-04
 ==================